Maddy

+++ main_app/models.py
@@ -14,7 +14,10 @@ CATEGORIES = (
 
 class UserInfo(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
-    budget = models.IntegerField()
+    budget = models.DecimalField(
+        max_digits=9,
+        decimal_places=2
+    )
 
 class Bill(models.Model):
     name = models.CharField(max_length=100)
